move lib inside of src
