+++ src/lib/InterfaceQt/moc_mainwindowphotoshop.cpp
@@ -1,0 +1,271 @@
+/****************************************************************************
+** Meta object code from reading C++ file 'mainwindowphotoshop.h'
+**
+** Created by: The Qt Meta Object Compiler version 67 (Qt 5.12.5)
+**
+** WARNING! All changes made in this file will be lost!
+*****************************************************************************/
+#include "mainwindowphotoshop.h"
+#include <QtCore/qbytearray.h>
+#include <QtCore/qmetatype.h>
+
+#if !defined(Q_MOC_OUTPUT_REVISION)
+#error "The header file 'mainwindowphotoshop.h' doesn't include <QObject>."
+#elif Q_MOC_OUTPUT_REVISION != 67
+#error "This file was generated using the moc from 5.12.5. It"
+#error "cannot be used with the include files from this version of Qt."
+#error "(The moc has changed too much.)"
+#endif
+
+QT_BEGIN_MOC_NAMESPACE
+QT_WARNING_PUSH
+QT_WARNING_DISABLE_DEPRECATED
+
+struct qt_meta_stringdata_MainWindowPhotoShop_t
+{
+	QByteArrayData data[23];
+	char stringdata0[768];
+};
+
+#define QT_MOC_LITERAL(idx, ofs, len) \
+ Q_STATIC_BYTE_ARRAY_DATA_HEADER_INITIALIZER_WITH_OFFSET(len, \
+ qptrdiff(offsetof(qt_meta_stringdata_MainWindowPhotoShop_t, stringdata0) + ofs \
+ - idx * sizeof(QByteArrayData)) \
+ )
+ 
+static const qt_meta_stringdata_MainWindowPhotoShop_t qt_meta_stringdata_MainWindowPhotoShop =
+{
+	{
+		QT_MOC_LITERAL(0, 0, 19), // "MainWindowPhotoShop"
+		QT_MOC_LITERAL(1, 20, 26), // "on_actionQuitter_triggered"
+		QT_MOC_LITERAL(2, 47, 0), // ""
+		QT_MOC_LITERAL(3, 48, 34), // "on_actionCharger_ImageNB_trig..."
+		QT_MOC_LITERAL(4, 83, 38), // "on_actionEnregistrer_ImageNB_..."
+		QT_MOC_LITERAL(5, 122, 35), // "on_actionCharger_ImageRGB_tri..."
+		QT_MOC_LITERAL(6, 158, 41), // "on_tableWidgetImages_itemSele..."
+		QT_MOC_LITERAL(7, 200, 32), // "on_pushButtonModifierNom_clicked"
+		QT_MOC_LITERAL(8, 233, 43), // "on_actionCouleur_TRUE_pour_Im..."
+		QT_MOC_LITERAL(9, 277, 44), // "on_actionCouleur_FALSE_pour_i..."
+		QT_MOC_LITERAL(10, 322, 30), // "on_pushButtonOperande1_clicked"
+		QT_MOC_LITERAL(11, 353, 38), // "on_pushButtonSupprimeOperande..."
+		QT_MOC_LITERAL(12, 392, 30), // "on_pushButtonOperande2_clicked"
+		QT_MOC_LITERAL(13, 423, 39), // "on_pushButtonSupprimerOperand..."
+		QT_MOC_LITERAL(14, 463, 29), // "on_pushButtonResultat_clicked"
+		QT_MOC_LITERAL(15, 493, 38), // "on_pushButtonSupprimerResulta..."
+		QT_MOC_LITERAL(16, 532, 31), // "on_pushButtonTraitement_clicked"
+		QT_MOC_LITERAL(17, 564, 39), // "on_actionEnregistrer_ImageRGB..."
+		QT_MOC_LITERAL(18, 604, 37), // "on_actionEnregistrer_ImageB_t..."
+		QT_MOC_LITERAL(19, 642, 37), // "on_actionImage_selectionn_e_t..."
+		QT_MOC_LITERAL(20, 680, 31), // "on_actionImage_par_id_triggered"
+		QT_MOC_LITERAL(21, 712, 30), // "on_actionImporterCSV_triggered"
+		QT_MOC_LITERAL(22, 743, 24) // "on_actionReset_triggered"
+	},
+	"MainWindowPhotoShop\0on_actionQuitter_triggered\0"
+	"\0on_actionCharger_ImageNB_triggered\0"
+	"on_actionEnregistrer_ImageNB_triggered\0"
+	"on_actionCharger_ImageRGB_triggered\0"
+	"on_tableWidgetImages_itemSelectionChanged\0"
+	"on_pushButtonModifierNom_clicked\0"
+	"on_actionCouleur_TRUE_pour_ImageB_triggered\0"
+	"on_actionCouleur_FALSE_pour_imageB_triggered\0"
+	"on_pushButtonOperande1_clicked\0"
+	"on_pushButtonSupprimeOperande1_clicked\0"
+	"on_pushButtonOperande2_clicked\0"
+	"on_pushButtonSupprimerOperande2_clicked\0"
+	"on_pushButtonResultat_clicked\0"
+	"on_pushButtonSupprimerResultat_clicked\0"
+	"on_pushButtonTraitement_clicked\0"
+	"on_actionEnregistrer_ImageRGB_triggered\0"
+	"on_actionEnregistrer_ImageB_triggered\0"
+	"on_actionImage_selectionn_e_triggered\0"
+	"on_actionImage_par_id_triggered\0"
+	"on_actionImporterCSV_triggered\0"
+	"on_actionReset_triggered"
+};
+#undef QT_MOC_LITERAL
+
+static const uint qt_meta_data_MainWindowPhotoShop[] =
+{
+// content:
+	8, // revision
+	0, // classname
+	0, 0, // classinfo
+	21, 14, // methods
+	0, 0, // properties
+	0, 0, // enums/sets
+	0, 0, // constructors
+	0, // flags
+	0, // signalCount
+
+// slots: name, argc, parameters, tag, flags
+	1, 0, 119, 2, 0x08 /* Private */,
+	3, 0, 120, 2, 0x08 /* Private */,
+	4, 0, 121, 2, 0x08 /* Private */,
+	5, 0, 122, 2, 0x08 /* Private */,
+	6, 0, 123, 2, 0x08 /* Private */,
+	7, 0, 124, 2, 0x08 /* Private */,
+	8, 0, 125, 2, 0x08 /* Private */,
+	9, 0, 126, 2, 0x08 /* Private */,
+	10, 0, 127, 2, 0x08 /* Private */,
+	11, 0, 128, 2, 0x08 /* Private */,
+	12, 0, 129, 2, 0x08 /* Private */,
+	13, 0, 130, 2, 0x08 /* Private */,
+	14, 0, 131, 2, 0x08 /* Private */,
+	15, 0, 132, 2, 0x08 /* Private */,
+	16, 0, 133, 2, 0x08 /* Private */,
+	17, 0, 134, 2, 0x08 /* Private */,
+	18, 0, 135, 2, 0x08 /* Private */,
+	19, 0, 136, 2, 0x08 /* Private */,
+	20, 0, 137, 2, 0x08 /* Private */,
+	21, 0, 138, 2, 0x08 /* Private */,
+	22, 0, 139, 2, 0x08 /* Private */,
+
+// slots: parameters
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+	QMetaType::Void,
+
+	0 // eod
+};
+
+void MainWindowPhotoShop::qt_static_metacall(QObject *_o, QMetaObject::Call _c, int _id, void **_a)
+{
+	if (_c == QMetaObject::InvokeMetaMethod)
+	{
+		auto *_t = static_cast<MainWindowPhotoShop *>(_o);
+		Q_UNUSED(_t)
+		switch (_id)
+		{
+		case 0:
+			_t->on_actionQuitter_triggered();
+			break;
+		case 1:
+			_t->on_actionCharger_ImageNB_triggered();
+			break;
+		case 2:
+			_t->on_actionEnregistrer_ImageNB_triggered();
+			break;
+		case 3:
+			_t->on_actionCharger_ImageRGB_triggered();
+			break;
+		case 4:
+			_t->on_tableWidgetImages_itemSelectionChanged();
+			break;
+		case 5:
+			_t->on_pushButtonModifierNom_clicked();
+			break;
+		case 6:
+			_t->on_actionCouleur_TRUE_pour_ImageB_triggered();
+			break;
+		case 7:
+			_t->on_actionCouleur_FALSE_pour_imageB_triggered();
+			break;
+		case 8:
+			_t->on_pushButtonOperande1_clicked();
+			break;
+		case 9:
+			_t->on_pushButtonSupprimeOperande1_clicked();
+			break;
+		case 10:
+			_t->on_pushButtonOperande2_clicked();
+			break;
+		case 11:
+			_t->on_pushButtonSupprimerOperande2_clicked();
+			break;
+		case 12:
+			_t->on_pushButtonResultat_clicked();
+			break;
+		case 13:
+			_t->on_pushButtonSupprimerResultat_clicked();
+			break;
+		case 14:
+			_t->on_pushButtonTraitement_clicked();
+			break;
+		case 15:
+			_t->on_actionEnregistrer_ImageRGB_triggered();
+			break;
+		case 16:
+			_t->on_actionEnregistrer_ImageB_triggered();
+			break;
+		case 17:
+			_t->on_actionImage_selectionn_e_triggered();
+			break;
+		case 18:
+			_t->on_actionImage_par_id_triggered();
+			break;
+		case 19:
+			_t->on_actionImporterCSV_triggered();
+			break;
+		case 20:
+			_t->on_actionReset_triggered();
+			break;
+		default:
+			;
+		}
+	}
+	Q_UNUSED(_a);
+}
+
+QT_INIT_METAOBJECT const QMetaObject MainWindowPhotoShop::staticMetaObject = { {
+		&QMainWindow::staticMetaObject,
+		qt_meta_stringdata_MainWindowPhotoShop.data,
+		qt_meta_data_MainWindowPhotoShop,
+		qt_static_metacall,
+		nullptr,
+		nullptr
+	}
+};
+
+const QMetaObject *MainWindowPhotoShop::metaObject() const
+{
+	return QObject::d_ptr->metaObject ? QObject::d_ptr->dynamicMetaObject() : &staticMetaObject;
+}
+
+void *MainWindowPhotoShop::qt_metacast(const char *_clname)
+{
+	if (!_clname) return nullptr;
+	if (!strcmp(_clname, qt_meta_stringdata_MainWindowPhotoShop.stringdata0))
+		return static_cast<void*>(this);
+	return QMainWindow::qt_metacast(_clname);
+}
+
+int MainWindowPhotoShop::qt_metacall(QMetaObject::Call _c, int _id, void **_a)
+{
+	_id = QMainWindow::qt_metacall(_c, _id, _a);
+	if (_id < 0)
+		return _id;
+	if (_c == QMetaObject::InvokeMetaMethod)
+	{
+		if (_id < 21)
+			qt_static_metacall(this, _c, _id, _a);
+		_id -= 21;
+	}
+	else if (_c == QMetaObject::RegisterMethodArgumentMetaType)
+	{
+		if (_id < 21)
+			*reinterpret_cast<int*>(_a[0]) = -1;
+		_id -= 21;
+	}
+	return _id;
+}
+
+QT_WARNING_POP
+QT_END_MOC_NAMESPACE
