+++ src/lib/images/bruitPoivreEtSel.bmp
