+++ src/lib/images/lena.bmp
